Fix buga w maksymalnym możliwym transferze do bufora(mierzonym w dł. najdłuższej kolejki ze stanu ACTIVE).

+++ tests/cpptest-1.1.2/test/mixer.h
@@ -13,7 +13,7 @@ public:
 	void mixer(mixer_input* inputs, size_t n, void* output_buf, 
 		size_t* output_size, unsigned long tx_interval_ms) const;
 private:
-	unsigned long get_total_bytes(const mixer_input* inputs, const int size) const;
+	size_t get_total_bytes(const mixer_input* inputs, const int size) const;
 
 	void init_consumed_bytes(mixer_input* inputs, const int size) const;
